+++ fme-android/src/main/java/com/vwo/VWOBuilder.kt
@@ -341,4 +341,5 @@ open class VWOBuilder(private val options: VWOInitOptions?) {
             mapOf("status" to status)
         )
     }
+
 }