+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.9.0] - 2025-09-10
+
+### Added
+- Support to use Device ID if `context.id` is not available
+
+```java
+// Java
+VWOUserContext context = new VWOUserContext();
+context.setId(""); // Empty ID to trigger device ID fallback
+context.setShouldUseDeviceIdAsUserId(true); // Use device ID as user ID
+```
+
+- Add support for user aliasing (will work after gateway has been setup)
+```java
+// Java
+VWOInitOptions options = new VWOInitOptions();
+options.isAliasingEnabled = true;
+
+VWOUserContext context = new VWOUserContext();
+context.id = "temp_id";
+VWO.setAlias(context, "user_alias");
+```
+
 ## [1.8.0] - 2025-08-29
 
 ### Added